pelna obsluga user

+++ Biblioteka/user.xaml.cs
@@ -39,21 +39,42 @@ namespace Biblioteka
             {
                 userDTOs.Add(UserTransformer.CreateDto(it));
             }
+            
             userTable.ItemsSource = userDTOs;
         }
 
         private void AddUserButton(object sender, RoutedEventArgs e)
         {
             AddUserWindow addUserWindow = new AddUserWindow();
-            addUserWindow.Show();
+            addUserWindow.ShowDialog();
             this.loadUsers();
         }
 
         private void RemoveUserButton(object sender, RoutedEventArgs e)
         {
             UserDTO userDto = (UserDTO)userTable.SelectedCells.FirstOrDefault().Item;
-            this.userService.removeUser(userDto.Id);
+            if (userDto != null)
+                this.userService.removeUser(userDto.Id);
+            else
+                MessageBox.Show("Aby usunąć użytkownika, wybierz go w tabeli!", "Usuwanie użytkownika",
+                    MessageBoxButton.OK, MessageBoxImage.Information);
             this.loadUsers();
         }
+
+        private void ChangeUser(object sender, RoutedEventArgs e)
+        {
+            UserDTO userDto = (UserDTO)userTable.SelectedCells.FirstOrDefault().Item;
+            if (userDto != null)
+            {
+                ChangeUserWindow changeUserWindow = new ChangeUserWindow();
+                changeUserWindow.initData(userDto);
+                changeUserWindow.ShowDialog();
+                this.loadUsers();
+            }
+            else
+                MessageBox.Show("Aby zmienić dane użytkownika, wybierz go w tabeli!", "Usuwanie użytkownika",
+                    MessageBoxButton.OK, MessageBoxImage.Information);
+        }
+
     }
 }
